Move `Context` to `stateful`
