+++ packages/engine/src/datastore/store.rs
@@ -1,9 +1,6 @@
-use stateful::state::State;
+use stateful::{context::Context, state::State};
 
-use crate::datastore::{
-    error::{Error, Result},
-    table::context::Context,
-};
+use crate::datastore::error::{Error, Result};
 
 /// The underlying state of a simulation run.
 pub struct Store {
