+++ packages/engine/src/datastore/table/task_shared_store.rs
@@ -1,14 +1,14 @@
 //! TODO: DOC
 use std::fmt::Debug;
 
-use stateful::state::{StateReadProxy, StateWriteProxy};
+use stateful::{
+    context::Context,
+    state::{StateReadProxy, StateWriteProxy},
+};
 
 use crate::{
     config::{StateBatchDistribution, Worker, WorkerAllocation},
-    datastore::{
-        error::{Error, Result},
-        table::context::Context,
-    },
+    datastore::error::{Error, Result},
     simulation::task::handler::SplitConfig,
 };
