+++ packages/engine/src/simulation/comms/mod.rs
@@ -26,7 +26,7 @@ pub mod package;
 
 use std::sync::{Arc, RwLock};
 
-use stateful::{agent::Agent, field::PackageId, state::StateReadProxy};
+use stateful::{agent::Agent, context::Context, field::PackageId, state::StateReadProxy};
 use uuid::Uuid;
 
 use self::message::{EngineToWorkerPoolMsg, WrappedTask};
@@ -37,7 +37,6 @@ use super::{
 };
 use crate::{
     datastore::table::{
-        context::Context,
         sync::{ContextBatchSync, StateSync, SyncPayload, WaitableStateSync},
         task_shared_store::TaskSharedStore,
     },
