+++ packages/engine/src/simulation/package/output/mod.rs
@@ -5,6 +5,7 @@ use std::sync::Arc;
 use async_trait::async_trait;
 pub use packages::{Name, OutputTask, OutputTaskMessage, PACKAGE_CREATORS};
 use stateful::{
+    context::Context,
     field::{FieldSpecMapAccessor, RootFieldSpec, RootFieldSpecCreator},
     globals::Globals,
     state::State,
@@ -14,7 +15,6 @@ use tracing::Span;
 use self::packages::Output;
 use crate::{
     config::{ExperimentConfig, SimRunConfig},
-    datastore::table::context::Context,
     simulation::{
         comms::package::PackageComms,
         package::{
