+++ packages/engine/src/simulation/package/state/mod.rs
@@ -5,6 +5,7 @@ use std::sync::Arc;
 use async_trait::async_trait;
 pub use packages::{Name, StateTask, StateTaskMessage, PACKAGE_CREATORS};
 use stateful::{
+    context::Context,
     field::{FieldSpecMapAccessor, RootFieldSpec, RootFieldSpecCreator},
     globals::Globals,
     state::State,
@@ -13,7 +14,6 @@ use tracing::Span;
 
 use crate::{
     config::{ExperimentConfig, SimRunConfig},
-    datastore::table::context::Context,
     simulation::{
         comms::package::PackageComms,
         package::{
